+++ app.js
@@ -1,5 +1,8 @@
 const express = require('express');
 const bodyParser= require ("body-parser");
+const { getDay } = require('./date');
+
+const date =require (__dirname + "/date.js")
 
 const app = express();
 const items=[];
@@ -12,20 +15,11 @@ app.use(express.static('public'));
 
 app.get('/', function(req, res) {
 
-    let today = new Date();
-    
-    let options = {
-        weekday:"long",
-        day: "numeric",
-        month:"long"
-    };
-
-    let day =today.toLocaleDateString("en-IN",options);
-    
-    
+    let day= date.getDate();
     res.render('list', {listTitle:day, newListItems:items});
 });
-    app.post("/",function (req,res){
+
+app.post("/",function (req,res){
 
     let item = req.body.newItem;
     
@@ -40,7 +34,8 @@ app.get('/', function(req, res) {
  });
  
  app.get("/work", function(req,res){
-     res.render("list",{listTitle:"Work list",newListItems:workItems});
+    let day= date.getDay();
+     res.render("list",{listTitle:day+" Work list",newListItems:workItems});
  })
